requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
PyJWT==2.6.0
mysql-connector-python==8.0.33
